Get FreqSpectrum instead of FreqResponse
* Added a function which generates the frequence spectrum from the given time
domain data using fft. If sampling rate is not given it is assumed as 44.1KHz.
Also a additional freqsubmult argument is given, this helps restrict the freq
to which the fft result is restricted.

Thus if no argument other than time domain data is given, it is assumed to be
a sampled at the popular audio frequency of 44.1Khz and inturn freq data upto
22.05Khz will be returned, because anything above 22.05Khz will not be sampled
in any meaningfull way at all with a sampling frequency of 44.1Khz (niquist
theorem).

Rather ideally the sampling rate should be atleast 10 times the maximum freq
to be properly monitored.

* Corrected the variable name from fnamegen to evalstr

* Fixed a bug wrt the logic which was generating the samples for a given bit
resolution. (2**bits-1) to ((2**bits-1)/2)

* Added logic to cross verify (reverify) that the generated data which is saved
into /tmp/data_.... files is proper by reloading and ploting the same wrt time
and freq domains.